Inject public path from the start script
We typically host our React apps within a backend app.  As a result,
the dev server will be in a different location than the window
location.  Because of this, we need the full public path, not just a
relative path. We know the public path in the `start` script, so we
inject it from there.