Switch to the stock webpack dev client (for now)
For now, we need to use the stock client because the custom dev client
relies on `window.location` to find the dev server.  However, since we
host our apps inside a back-end app, the dev server is running
somewhere else.

See https://github.com/CodingZeal/create-react-app/issues/6 for some
ideas on how to go back to the custom dev client.