Allow testing of CSS classes on components
By default, CRA returns an empty object when importing CSS files in
tests, resulting in every imported style being `null`.

We replace that with identity-obj-proxy, which causes every key in the
imported object to return its own name as the value.  This allows us to
test CSS styles properly.

To use the proxy, we have to replace the `transform` with
`moduleNameMapper`.