Improve allowing imports from application root path
CRA uses the `fallback:` setting in the webpack config files, which is
a better choice.  We splice our `appSrc` path into the list.

For Jest, we add our `appSrc` path to the default value for
`moduleDirectories`.  This needs to be tested, because the default
value may not actually work correctly.