Merge tag 'v1.0.7' into zeal
This merge is much more complicated than it should have been.  Apparently, the parent project spun off a 0.9.x branch around 0.9.2 or so and never merged it back into master.  Thus, there are a signficant number of changes here that resulted from conflicts between their branches.

In this commit, I modified our `lint` script to match the changes made upstream in the other scripts.

This commit merges some of our customizations in, but not all yet.  I'll add those back in later commits.

In particular:
* I haven't added Sass and CSS modules support back into the Webpack configuration files.

* I haven't added back the changes needed to resolve imports from the `client` directory.

* I haven't added back the changes needed to configure the public path in dev mode; the `start` script was refactored heavily, so we need to figure out how to do what we need to do.

* I haven't added back the changes needed to keep the browser from automatically opening.