Re-enable hosting inside a back-end application
We often host React apps in a back-end application.  In order for that to work, we need to customize the configuration in several places.

In 1.0, we turned the webpack dev configuration into a function that took `publicPath` as an argument.  In this version of CRA, the dev config is imported in more than one place, and not all places have the necessary information.

So, instead, we now duplicate the code for computing the public path.  This is unfortunate, but does work.

This version also makes it so that we can open the browser to the correct location (if the host application sets the `APP_PORT` env var appropriately), so we re-enable that code.